RDA3/FLASK_MySQL/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
PyMySQL
